@@ -2,26 +2,20 @@ package com.iti.sakila.persistance.entity;
 // Generated Apr 1, 2023, 5:10:50 AM by Hibernate Tools 6.1.7.Final
 
 
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.*;
+import org.mapstruct.ap.shaded.freemarker.template.SimpleDate;
+
+import javax.swing.text.DateFormatter;
 
 import static jakarta.persistence.GenerationType.IDENTITY;
 
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.OneToMany;
-import jakarta.persistence.Table;
-import jakarta.persistence.Temporal;
-import jakarta.persistence.TemporalType;
-
 import java.math.BigDecimal;
-import java.util.Date;
+import java.text.DateFormat;
+import java.text.ParseException;
+import java.text.SimpleDateFormat;
+import java.time.Year;
+import java.util.*;
 import java.sql.Timestamp;
-import java.util.HashSet;
-import java.util.Set;
 
 /**
  * Film generated by hbm2java
@@ -36,14 +30,14 @@ public class Film implements java.io.Serializable {
     private Language languageByOriginalLanguageId;
     private String title;
     private String description;
-    private Date releaseYear;
+    private int releaseYear;
     private byte rentalDuration;
     private BigDecimal rentalRate;
     private Short length;
     private BigDecimal replacementCost;
     private String rating;
     private String specialFeatures;
-    private Timestamp lastUpdate;
+    private Timestamp lastUpdate = new Timestamp(new Date().getTime());
     private Set<Inventory> inventories = new HashSet<Inventory>(0);
     private Set<FilmActor> filmActors = new HashSet<FilmActor>(0);
     private Set<FilmCategory> filmCategories = new HashSet<FilmCategory>(0);
@@ -61,7 +55,7 @@ public class Film implements java.io.Serializable {
         this.lastUpdate = lastUpdate;
     }
 
-    public Film(Language languageByLanguageId, Language languageByOriginalLanguageId, String title, String description, Date releaseYear, byte rentalDuration, BigDecimal rentalRate, Short length, BigDecimal replacementCost, String rating, String specialFeatures, Timestamp lastUpdate, Set<Inventory> inventories, Set<FilmActor> filmActors, Set<FilmCategory> filmCategories) {
+    public Film(Language languageByLanguageId, Language languageByOriginalLanguageId, String title, String description, int releaseYear, byte rentalDuration, BigDecimal rentalRate, Short length, BigDecimal replacementCost, String rating, String specialFeatures, Timestamp lastUpdate, Set<Inventory> inventories, Set<FilmActor> filmActors, Set<FilmCategory> filmCategories) {
         this.languageByLanguageId = languageByLanguageId;
         this.languageByOriginalLanguageId = languageByOriginalLanguageId;
         this.title = title;
@@ -81,8 +75,6 @@ public class Film implements java.io.Serializable {
 
     @Id
     @GeneratedValue(strategy = IDENTITY)
-
-
     @Column(name = "film_id", unique = true, nullable = false)
     public Short getFilmId() {
         return this.filmId;
@@ -92,7 +84,7 @@ public class Film implements java.io.Serializable {
         this.filmId = filmId;
     }
 
-    @ManyToOne(fetch = FetchType.EAGER)
+    @ManyToOne(fetch = FetchType.EAGER, cascade = CascadeType.ALL)
     @JoinColumn(name = "language_id", nullable = false)
     public Language getLanguageByLanguageId() {
         return this.languageByLanguageId;
@@ -134,11 +126,11 @@ public class Film implements java.io.Serializable {
 
     @Temporal(TemporalType.DATE)
     @Column(name = "release_year", length = 4)
-    public Date getReleaseYear() {
+    public int getReleaseYear() {
         return this.releaseYear;
     }
 
-    public void setReleaseYear(Date releaseYear) {
+    public void setReleaseYear(int releaseYear) {
         this.releaseYear = releaseYear;
     }
 
@@ -221,7 +213,7 @@ public class Film implements java.io.Serializable {
         this.inventories = inventories;
     }
 
-    @OneToMany(fetch = FetchType.EAGER, mappedBy = "film")
+    @OneToMany(fetch = FetchType.LAZY, mappedBy = "film")
     public Set<FilmActor> getFilmActors() {
         return this.filmActors;
     }
@@ -255,6 +247,8 @@ public class Film implements java.io.Serializable {
                 ", lastUpdate=" + lastUpdate +
                 '}';
     }
+
+
 }
 
 
