@@ -7,8 +7,7 @@ import jakarta.persistence.*;
 import static jakarta.persistence.GenerationType.IDENTITY;
 
 import java.sql.Timestamp;
-import java.util.HashSet;
-import java.util.Set;
+import java.util.*;
 
 /**
  * Actor generated by hbm2java
@@ -21,8 +20,8 @@ public class Actor implements java.io.Serializable {
     private Short actorId;
     private String firstName;
     private String lastName;
-    private Timestamp lastUpdate;
-    private Set<FilmActor> filmActors = new HashSet<FilmActor>(0);
+    private Timestamp lastUpdate = new Timestamp(new Date().getTime());
+    private Set<FilmActor> filmActors = new HashSet<>(0);
 
     public Actor() {
     }
@@ -84,7 +83,7 @@ public class Actor implements java.io.Serializable {
         this.lastUpdate = lastUpdate;
     }
 
-    @OneToMany(fetch = FetchType.LAZY, mappedBy = "actor", cascade = CascadeType.ALL)
+    @OneToMany(fetch = FetchType.EAGER, mappedBy = "actor", cascade = CascadeType.ALL)
     public Set<FilmActor> getFilmActors() {
         return this.filmActors;
     }
