@@ -2,22 +2,17 @@ package com.iti.sakila.persistance.entity;
 // Generated Apr 1, 2023, 5:10:50 AM by Hibernate Tools 6.1.7.Final
 
 
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.*;
+
+import org.hibernate.annotations.Type;
+
+
 
 import static jakarta.persistence.GenerationType.IDENTITY;
 
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.OneToMany;
-import jakarta.persistence.Table;
-import jakarta.persistence.Temporal;
-import jakarta.persistence.TemporalType;
-
+import java.awt.*;
 import java.sql.Timestamp;
+import java.util.Date;
 import java.util.HashSet;
 import java.util.Set;
 
@@ -25,9 +20,9 @@ import java.util.Set;
  * Address generated by hbm2java
  */
 @Entity
-@Table(name = "address"
-        , catalog = "sakila"
-)
+@Table(name = "address", catalog = "sakila", indexes = {
+        @Index(name = "idx_address_location", columnList = "location")
+})
 public class Address implements java.io.Serializable {
     private Short addressId;
     private City city;
@@ -36,8 +31,8 @@ public class Address implements java.io.Serializable {
     private String district;
     private String postalCode;
     private String phone;
-    private byte[] location;
-    private Timestamp lastUpdate;
+    private Object location;
+    private Timestamp lastUpdate = new Timestamp(new Date().getTime());
     private Set<Staff> staffs = new HashSet<Staff>(0);
     private Set<Customer> customers = new HashSet<Customer>(0);
     private Set<Store> stores = new HashSet<Store>(0);
@@ -46,7 +41,7 @@ public class Address implements java.io.Serializable {
     }
 
 
-    public Address(City city, String address, String district, String phone, byte[] location, Timestamp lastUpdate) {
+    public Address(City city, String address, String district, String phone, Object location, Timestamp lastUpdate) {
         this.city = city;
         this.address = address;
         this.district = district;
@@ -55,7 +50,7 @@ public class Address implements java.io.Serializable {
         this.lastUpdate = lastUpdate;
     }
 
-    public Address(City city, String address, String address2, String district, String postalCode, String phone, byte[] location, Timestamp lastUpdate, Set<Staff> staffs, Set<Customer> customers, Set<Store> stores) {
+    public Address(City city, String address, String address2, String district, String postalCode, String phone, Object location, Timestamp lastUpdate, Set<Staff> staffs, Set<Customer> customers, Set<Store> stores) {
         this.city = city;
         this.address = address;
         this.address2 = address2;
@@ -71,8 +66,6 @@ public class Address implements java.io.Serializable {
 
     @Id
     @GeneratedValue(strategy = IDENTITY)
-
-
     @Column(name = "address_id", unique = true, nullable = false)
     public Short getAddressId() {
         return this.addressId;
@@ -82,7 +75,7 @@ public class Address implements java.io.Serializable {
         this.addressId = addressId;
     }
 
-    @ManyToOne(fetch = FetchType.LAZY)
+    @ManyToOne(fetch = FetchType.LAZY, cascade = CascadeType.ALL)
     @JoinColumn(name = "city_id", nullable = false)
     public City getCity() {
         return this.city;
@@ -142,13 +135,18 @@ public class Address implements java.io.Serializable {
         this.phone = phone;
     }
 
+//    TODO [JPA Buddy] create field to map the 'location' column
+//    Available actions: Define target Java type | Uncomment as is | Remove column mapping
+//    @Column(name = "location", columnDefinition = "GEOMETRY(65535) not null")
+//    private Object location;
 
-    @Column(name = "location", nullable = false)
-    public byte[] getLocation() {
+    @Column(name = "location", columnDefinition = "GEOMETRY(65535) not null")
+    public Object getLocation()
+    {
         return this.location;
     }
 
-    public void setLocation(byte[] location) {
+    public void setLocation(Object location) {
         this.location = location;
     }
 
