@@ -13,8 +13,7 @@ import jakarta.persistence.Table;
 import jakarta.persistence.Temporal;
 import jakarta.persistence.TemporalType;
 import java.sql.Timestamp;
-import java.util.ArrayList;
-import java.util.List;
+import java.util.*;
 
 /**
  * Category generated by hbm2java
@@ -24,10 +23,10 @@ import java.util.List;
     ,catalog="sakila"
 )
 public class Category  implements java.io.Serializable {
-     private Byte categoryId;
-     private String name;
-     private Timestamp lastUpdate;
-     private List<FilmCategory> filmCategories = new ArrayList<>(0);
+    private int categoryId;
+    private String name;
+    private Timestamp lastUpdate = new Timestamp(new Date().getTime());
+     private Set<FilmCategory> filmCategories = new HashSet<>(0);
 
     public Category() {
     }
@@ -37,7 +36,7 @@ public class Category  implements java.io.Serializable {
         this.name = name;
         this.lastUpdate = lastUpdate;
     }
-    public Category(String name, Timestamp lastUpdate, List<FilmCategory> filmCategories) {
+    public Category(String name, Timestamp lastUpdate, Set<FilmCategory> filmCategories) {
        this.name = name;
        this.lastUpdate = lastUpdate;
        this.filmCategories = filmCategories;
@@ -47,11 +46,11 @@ public class Category  implements java.io.Serializable {
 
     
     @Column(name="category_id", unique=true, nullable=false)
-    public Byte getCategoryId() {
+    public int getCategoryId() {
         return this.categoryId;
     }
     
-    public void setCategoryId(Byte categoryId) {
+    public void setCategoryId(int categoryId) {
         this.categoryId = categoryId;
     }
 
@@ -76,11 +75,11 @@ public class Category  implements java.io.Serializable {
     }
 
 @OneToMany(fetch=FetchType.LAZY, mappedBy="category")
-    public List<FilmCategory> getFilmCategories() {
+    public Set<FilmCategory> getFilmCategories() {
         return this.filmCategories;
     }
     
-    public void setFilmCategories(List<FilmCategory> filmCategories) {
+    public void setFilmCategories(Set<FilmCategory> filmCategories) {
         this.filmCategories = filmCategories;
     }
 
