@@ -13,80 +13,85 @@ import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import jakarta.persistence.Temporal;
 import jakarta.persistence.TemporalType;
+
 import java.sql.Timestamp;
+import java.util.Date;
 
 /**
  * FilmActor generated by hbm2java
  */
 @Entity
-@Table(name="film_actor"
-    ,catalog="sakila"
+@Table(name = "film_actor"
+        , catalog = "sakila"
 )
-public class FilmActor  implements java.io.Serializable {
-
-
-     private FilmActorId id;
-     private Film film;
-     private Actor actor;
-     private Timestamp lastUpdate;
+public class FilmActor implements java.io.Serializable {
+    private FilmActorId id;
+    private Film film;
+    private Actor actor;
+    private Timestamp lastUpdate = new Timestamp(new Date().getTime());
 
     public FilmActor() {
     }
 
     public FilmActor(FilmActorId id, Film film, Actor actor, Timestamp lastUpdate) {
-       this.id = id;
-       this.film = film;
-       this.actor = actor;
-       this.lastUpdate = lastUpdate;
+        this.id = id;
+        this.film = film;
+        this.actor = actor;
+        this.lastUpdate = lastUpdate;
     }
-   
-     @EmbeddedId
 
-    
-    @AttributeOverrides( {
-        @AttributeOverride(name="actorId", column=@Column(name="actor_id", nullable=false) ), 
-        @AttributeOverride(name="filmId", column=@Column(name="film_id", nullable=false) ) } )
+    @EmbeddedId
+    @AttributeOverrides({
+            @AttributeOverride(name = "actorId", column = @Column(name = "actor_id", nullable = false)),
+            @AttributeOverride(name = "filmId", column = @Column(name = "film_id", nullable = false))})
     public FilmActorId getId() {
         return this.id;
     }
-    
+
     public void setId(FilmActorId id) {
         this.id = id;
     }
 
-@ManyToOne(fetch=FetchType.LAZY)
-    @JoinColumn(name="film_id", nullable=false, insertable=false, updatable=false)
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "film_id", nullable = false, insertable = false, updatable = false)
     public Film getFilm() {
         return this.film;
     }
-    
+
     public void setFilm(Film film) {
         this.film = film;
     }
 
-@ManyToOne(fetch=FetchType.LAZY)
-    @JoinColumn(name="actor_id", nullable=false, insertable=false, updatable=false)
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "actor_id", nullable = false, insertable = false, updatable = false)
     public Actor getActor() {
         return this.actor;
     }
-    
+
     public void setActor(Actor actor) {
         this.actor = actor;
     }
 
     @Temporal(TemporalType.TIMESTAMP)
-    @Column(name="last_update", nullable=false, length=19)
+    @Column(name = "last_update", nullable = false, length = 19)
     public Timestamp getLastUpdate() {
         return this.lastUpdate;
     }
-    
+
     public void setLastUpdate(Timestamp lastUpdate) {
         this.lastUpdate = lastUpdate;
     }
 
 
-
-
+    @Override
+    public String toString() {
+        return "FilmActor{" +
+                "id=" + id +
+                ", film=" + film +
+                ", actor=" + actor +
+                ", lastUpdate=" + lastUpdate +
+                '}';
+    }
 }
 
 
