@@ -14,6 +14,7 @@ import jakarta.persistence.Table;
 import jakarta.persistence.Temporal;
 import jakarta.persistence.TemporalType;
 import java.sql.Timestamp;
+import java.util.Date;
 
 /**
  * FilmCategory generated by hbm2java
@@ -26,7 +27,7 @@ public class FilmCategory  implements java.io.Serializable {
      private FilmCategoryId id;
      private Film film;
      private Category category;
-     private Timestamp lastUpdate;
+    private Timestamp lastUpdate =new Timestamp(new Date().getTime());
 
     public FilmCategory() {
     }
