@@ -1,10 +1,19 @@
 package com.iti.sakila.utils;
 
-import java.math.BigDecimal;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
 
-public record FilterRecord(int page,
-                           int categoryId,
-                           BigDecimal price,
-                           BigDecimal rate,
-                           String name) {
+import java.math.BigDecimal;
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+public final class FilterRecord {
+    private  int page = 1;
+    private  int categoryId = 0;
+    private  BigDecimal price = new BigDecimal(0);
+    private  BigDecimal rate = new BigDecimal(0);
+    private  String name = "";
 }
