@@ -1,20 +1,23 @@
 package com.iti.sakila.service;
 
-import com.iti.sakila.bussiness.dtos.ActorDto;
-import com.iti.sakila.bussiness.dtos.ListResponse;
+import com.iti.sakila.bussiness.dtos.actordto.ActorDto;
+import com.iti.sakila.bussiness.mapper.actormapper.ActorMapper;
 import com.iti.sakila.bussiness.service.ActorService;
-import com.sun.xml.wss.saml.assertion.saml11.jaxb20.Assertion;
+import com.iti.sakila.bussiness.service.BaseService;
+import com.iti.sakila.persistance.entity.Actor;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 
+import java.util.List;
+
 public class ActorServiceTest {
 
-    public static ActorService actorService;
+    private static BaseService<Actor,ActorDto> actorService;
 
     @BeforeAll
     public static void init() {
-        actorService = new ActorService();
+        actorService = new ActorService(Actor.class, ActorMapper.INSTANCE);
     }
 
     @Test
@@ -22,9 +25,9 @@ public class ActorServiceTest {
         //Arrange
         int page = 1;
         //Act
-        ListResponse<ActorDto> response = actorService.getActors(page);
-        System.out.println(response.getItems());
+        List<ActorDto> response = actorService.getAll(page);
+        System.out.println(response);
         //Assert
-        Assertions.assertEquals(response.getMessage().getCode(), 200);
+        Assertions.assertEquals(response.size(), 10);
     }
 }
